refactor: split member price group form
