+++ src/app/[locale]/(back-office)/team/[teamId]/price-groups/_components/general-group/price-groups/price-group-form.tsx
@@ -1,0 +1,10 @@
+'use client'
+
+import { BasePriceGroupForm } from '../../shared/price-group-form'
+import type { BasePriceGroupFormProps } from '../../shared/price-group-form'
+
+export type GeneralPriceGroupFormProps = Omit<BasePriceGroupFormProps, 'statusType'>
+
+export default function GeneralPriceGroupForm(props: GeneralPriceGroupFormProps) {
+  return <BasePriceGroupForm {...props} statusType="GENERAL" />
+}
