+++ src/app/[locale]/(back-office)/team/[teamId]/price-groups/_components/member-group/index.ts
@@ -1,13 +1,14 @@
 export { BasicInfoForm } from '@/app/[locale]/(back-office)/team/[teamId]/chargers/_components/chargers/basic-info-form'
 export { OcppIntegrationForm } from '@/app/[locale]/(back-office)/team/[teamId]/chargers/_components/chargers/ocpp-integration-form'
 export { ConnectorForm } from '@/app/[locale]/(back-office)/team/[teamId]/chargers/_components/connectors/connector-form'
-export { PriceGroupForm } from './price-groups'
+export { MemberPriceGroupForm } from './price-groups'
 export type {
   FeeFormData,
   FormData,
   Mode,
   PriceFormData,
-  PriceGroupFormProps,
+  PriceGroupFormInitialData,
+  PriceGroupFormSubmitData,
   PriceType,
   StatusType,
 } from './price-groups'
