+++ src/app/[locale]/(back-office)/team/[teamId]/price-groups/_components/shared/price-group-form/index.ts
@@ -1,0 +1,12 @@
+export { BasePriceGroupForm } from './base-price-group-form'
+export type {
+  BasePriceGroupFormProps,
+  FeeFormData,
+  FormData,
+  Mode,
+  PriceFormData,
+  PriceGroupFormInitialData,
+  PriceGroupFormSubmitData,
+  PriceType,
+  StatusType,
+} from './types'
